Refactor filters and graphs as 2 separates entities
